refactor: split terminal input handling and share lifecycle fixture
