+++ docs/relay-validation-2026-08-30.md
@@ -24,7 +24,9 @@ The one-hour run completed 90,424 matching responses on eight clients, with 40 r
 
 The one-hour run uses the renewal implementation captured when that run started. Additional error-path hardening was reviewed and tested separately afterward; the final cross-language short/adversarial runs validate those later changes. The final actor-handler extraction only relocates the dispatch code and is covered by the Rust workspace suite. The fixture uses real cryptography and WebSocket transports but an echo actor instead of the full workbench. It cannot certify UI draft preservation, duplicate side effects in real commands, operating-system suspension or radio handoff.
 
-The scoped file-size ratchet passes after extracting relay handlers from `server.rs`. The optional repository-wide scan still flags two unchanged files already exceeding 500 lines in the starting revision: `mobile/lib/src/features/terminal/application/terminal_session_controller.dart` (501) and `mobile/test/host_connection_controller_test.dart` (507). Their baselines were not relaxed.
+The scoped file-size ratchet passed after extracting relay handlers from `server.rs`, but PR job `99205679849` exposed that CI requires the repository-wide scan too. That scan rejected two files already above 500 lines in the starting revision. The follow-up fix separates terminal input methods from `terminal_session_controller.dart` and shares the lifecycle fixture between connection and terminal tests. It preserves the test cases and the original size limit; no baseline exceptions were added.
+
+After this follow-up, the exact global file-size check, desktop formatting and analysis, mobile analysis and all 594 ordinary mobile tests passed locally. Riverpod was regenerated once for the extraction. The controller now has 456 lines and the connection test file has 495 lines. The failed remote job has not been rerun with these uncommitted changes.
 
 No Android device was attached according to `adb devices -l`. iOS device tooling is unavailable on this Linux host. Real Android/iOS tests remain a release gate: cold start, rapid foreground/background changes, file picker, Wi-Fi/cellular transition, suspension beyond 120 seconds and runtime restart. Use the same devices, builds, payloads and network shaping for a before/after p50/p95 and peak-memory comparison. No such baseline comparison has been claimed here.
