+++ mobile/test/support/controlled_app_lifecycle.dart
@@ -1,0 +1,15 @@
+import 'package:alera_mobile/src/app/lifecycle/app_lifecycle_controller.dart';
+import 'package:flutter/widgets.dart';
+
+class ControlledAppLifecycle extends AppLifecycleController {
+  ControlledAppLifecycle([this.initialState = AppLifecycleState.paused]);
+
+  final AppLifecycleState initialState;
+
+  @override
+  AppLifecycleState build() => initialState;
+
+  void setLifecycleState(AppLifecycleState next) {
+    state = next;
+  }
+}
